% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ball mapper.R
\name{ball_mapper}
\alias{ball_mapper}
\title{Ball mapper}
\usage{
ball_mapper(D, epsilon, id = NULL)
}
\arguments{
\item{D}{A matrix of distances.}

\item{epsilon}{A non-negative number.}

\item{id}{A possible subset of indexes of D.}
}
\description{
Ball mapper
}
