% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.dbscan}
\alias{clust.dbscan}
\title{DBScan clustering function}
\usage{
clust.dbscan(
  distance_matrix,
  epsilon = 1,
  min_points_per_cluster = 1,
  border_points = TRUE,
  put_all_outliers_in_a_single_cluster = FALSE,
  ...
)
}
\description{
DBScan clustering function
}
