% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.first_cutoff}
\alias{clust.first_cutoff}
\title{First cutoff clustering function}
\usage{
clust.first_cutoff(
  distance_matrix,
  num_breaks_in_histogram_of_distances = 10,
  ...
)
}
\description{
First cutoff clustering function
}
