% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clust.hdbscan}
\alias{clust.hdbscan}
\title{HDBScan clustering function}
\usage{
clust.hdbscan(
  distance_matrix,
  n_points_dist = NULL,
  min_pts_per_cluster = NULL,
  min_percent_points_per_cluster = 5,
  put_all_outliers_in_a_single_cluster = FALSE
)
}
\description{
HDBScan clustering function
}
