% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate data.R
\name{data.noisy_circle}
\alias{data.noisy_circle}
\title{Noisy circle data}
\usage{
data.noisy_circle(n = 1000, radius = 1, sd = 0.1)
}
\arguments{
\item{n}{The amount of points in the sample.}

\item{radius}{The radius of the circle.}

\item{sd}{The standard deviation.}
}
\value{
A tibble.
}
\description{
Noisy circle data
}
\examples{
X = data.noisy_circle()
plot(X)
}
