% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_point_to_set}
\alias{dist_point_to_set}
\title{Distance from a point x to a set X}
\usage{
dist_point_to_set(x, X)
}
\arguments{
\item{x}{A point.}

\item{X}{A vector of points.}
}
\value{
A vector of the same size as x, with the corresponding distance from x to each element of X.
}
\description{
Distance from a point x to a set X
}
