% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{farthest_points_sample}
\alias{farthest_points_sample}
\title{Farthest point sample of a set}
\usage{
farthest_points_sample(X, p = 1000)
}
\arguments{
\item{X}{A dataframe or matrix of points.}

\item{p}{The amount of points after the procedure.}
}
\description{
Farthest point sample of a set
}
