% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{mapper}
\alias{mapper}
\title{Calculate the mapper graph of a set}
\usage{
mapper(
  X = data.noisy_circle(),
  distance = stats::dist,
  f_X = X[1, ],
  covering = cover.uniform,
  clustering_function = clust.first_cutoff
)
}
\arguments{
\item{X}{A dataframe or matrix.}

\item{distance}{A dist object or matrix of distances; or a function that return such object when applied to X.}

\item{f_X}{A numeric vector with length equal to the rows of X; or a function that return such a vector when applied to X.}

\item{covering}{A matrix with intervals or a function that returns such an matrix.}

\item{clustering_function}{A clustering function.}
}
\value{
A mapper object.
}
\description{
Calculate the mapper graph of a set
}
\examples{
mp = mapper()
mp \%>\% plot_mapper()
}
