% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{split_pullback}
\alias{split_pullback}
\title{Split the pullback of a covering}
\usage{
split_pullback(
  pullback,
  X_num = NULL,
  data = NULL,
  X = NULL,
  distance_matrix = NULL,
  distance_function = NULL,
  clustering_function = clustering_hdbscan
)
}
\description{
Split the pullback of a covering
}
