% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ball mapper.R
\name{ball_mapper}
\alias{ball_mapper}
\title{Ball mapper}
\usage{
ball_mapper(X, distance_matrix, epsilon, id = NULL)
}
\arguments{
\item{distance_matrix}{A matrix of distances.}

\item{epsilon}{A non-negative number.}

\item{id}{A possible subset of indexes of distance_matrix.}
}
\description{
Ball mapper
}
