% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mapper}
\alias{plot_mapper}
\title{Plot the mapper graph interactively}
\usage{
plot_mapper(
  mp,
  data_column = names(mp$data)[1],
  color_vertex_by = f,
  color_scale = viridis::viridis(10)
)
}
\description{
Plot the mapper graph interactively
}
